requirements.txt
# Web框架
flask>=3.0.0

# 数据库
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Redis
redis>=5.0.0

# 任务队列
celery>=5.3.0

# 数据验证
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 爬虫
crawl4ai>=0.2.0
beautifulsoup4>=4.12.0

# AI
zhipuai>=2.0.0

# Web服务器
gunicorn>=21.0.0

# 工具
python-dotenv>=1.0.0

# HTML解析
lxml>=5.0.0
